Continued Kindle stuff
This code gets messier with every update. Sorry.

+++ src/at/yawk/kindle/Kindle.java
@@ -14,10 +14,11 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 public class Kindle {
-	private final File				bookRoot;
-	private final File				kindleRoot;
+	private final File					bookRoot;
+	private final File					kindleRoot;
 	
-	private final Map<String, Long>	specialModifyDate	= new HashMap<String, Long>();
+	private final Map<String, Long>		specialModifyDate	= new HashMap<>();
+	private final Map<String, String>	realName			= new HashMap<>();
 	
 	Kindle(final File kindleRootDir) {
 		bookRoot = new File(kindleRoot = kindleRootDir, "documents");
@@ -25,7 +26,9 @@ public class Kindle {
 			try {
 				final DataInputStream dis = new DataInputStream(new FileInputStream(getSpecialModifyDateFile()));
 				while(dis.available() > 0) {
-					specialModifyDate.put(dis.readUTF(), dis.readLong());
+					final String s = dis.readUTF();
+					realName.put(s, dis.readUTF());
+					specialModifyDate.put(s, dis.readLong());
 				}
 				dis.close();
 			} catch(IOException e) {
@@ -43,12 +46,15 @@ public class Kindle {
 	}
 	
 	public void saveSpecialDates() throws IOException {
-		final DataOutputStream dos = new DataOutputStream(new FileOutputStream(getSpecialModifyDateFile()));
-		for(final Entry<String, Long> e : specialModifyDate.entrySet()) {
-			dos.writeUTF(e.getKey());
-			dos.writeLong(e.getValue());
+		if(!specialModifyDate.isEmpty() || getSpecialModifyDateFile().exists()) {
+			final DataOutputStream dos = new DataOutputStream(new FileOutputStream(getSpecialModifyDateFile()));
+			for(final Entry<String, Long> e : specialModifyDate.entrySet()) {
+				dos.writeUTF(e.getKey());
+				dos.writeUTF(realName.get(e.getKey()));
+				dos.writeLong(e.getValue());
+			}
+			dos.close();
 		}
-		dos.close();
 	}
 	
 	public boolean checkOnKindle(final String bookName) {
@@ -69,8 +75,9 @@ public class Kindle {
 		return null;
 	}
 	
-	public void setSpecialModifyDate(final String bookName, final long modifyDate) {
+	public void setSpecialModifyDate(final String bookName, final String realBookName, final long modifyDate) {
 		specialModifyDate.put(bookName, modifyDate);
+		realName.put(bookName, realBookName);
 	}
 	
 	public void deleteFromKindle(final String bookName) {
@@ -92,8 +99,21 @@ public class Kindle {
 		})[0].lastModified();
 	}
 	
+	public Long getSpecialBookUpdate(final String bookName) {
+		return specialModifyDate.get(bookName);
+	}
+	
+	public long getBookUpdateDateTotal(final String bookName) {
+		final Long l = getSpecialBookUpdate(bookName);
+		return l == null ? getBookUpdateDateMilliseconds(bookName) : l;
+	}
+	
 	private final File getSpecialModifyDateFile() {
 		return new File(kindleRoot, "at.yawk.kindle.specialModifyDate.dat");
+	}
+	
+	public String getRealBookName(final String bookName) {
+		return realName.get(bookName);
 	}
 	
 	@Override
@@ -104,5 +124,18 @@ public class Kindle {
 	
 	public File getKindleRoot() {
 		return kindleRoot;
+	}
+	
+	public String[] listBooks() {
+		final String[] list = bookRoot.list(new FilenameFilter() {
+			@Override
+			public boolean accept(File arg0, String arg1) {
+				return arg1.matches("[a-zA-Z ]+-.+\\.azw");
+			}
+		});
+		for(int i = 0; i < list.length; i++) {
+			list[i] = list[i].substring(0, list[i].indexOf('-'));
+		}
+		return list;
 	}
 }
